Make color transitions sequence accept an iterator
I wanted to make the color transitions accept an iterator so that I
could do repeating color transition sequences, that were either simple
(like a breathing pulse) or as complex as I wanted.  This also allows
finite sequences that can be computed on the fly, instead of having to
populate a list and then passing that into a function to handle the
sequence.

One thing I learned the hard way: You can connect as many callbacks to
the a QPropertyAnimation.finish signal (or probably most QT signals) as
you like.  (This is intuitive, but escaped me in this case.)  I kept
adding the same on-finish callback in the `transition_to_color` method,
which called `handle_next_transition`.  This worked fine the first and
second time, but then called the callback two times simultaneously
(since the function had been called twice so far), and then four times,
then eight, and so on, literally in an exponential fashion.  (This was
not my intended behavior, if you weren't aware.)  I spent half a day
debugging this, only to realize what was going on when I was about to
give up in desperation.  When I added the needed and necessary
`self.color_animation.finished.disconnect()` (after taking a break from
everything), it worked, and progress suddenly became much more possible!
I'm happy that I was able to figure that out.  I hope I'm able to hold
on to that simple lesson in the future.

Whoo!