Make GlowBox work with the state machine
There are actually a lot of things that are going on here:

- Add descriptions to States and Events
- Split notification states into 2
- Use QTimer instead of scheduler
- Adjust line spacing on transitions
- Update logging messages
- Set glow notification to happen on transition to notification
- Allow specifying an easing curve for color transitions
- Make rounding to nearest even optional in
  `intervals_decreasing_over_total_time()`
- Refactor variable names to make more sense in
  `intervals_decreasing_over_total_time()`

I really need to be better about commiting after making changes that fit
well together.  (It's easy to squash commits if I end up with too
many; figuring out which changes aren't related after the fact though is
harder to do.)

One problem I ran into, when I was setting the glow box to follow an
iterable of transitions inside the main program, it worked fine, but
when I tried to do the exact same thing in the state machine, it would
hang.  The issue wasn't the state machine itself, but that a different
thread was processing the event that lead to the state in which the
pulsing was supposed to happen.  It turns out that QT (and hence PySide)
does not work well with the `threading` module in the Python standard
library.  Luckily, QT comes with its own set of tools to handle threaded
stuff.  For what I need so far, QTimer works just fine, but for running
`pynput` in a separate thread, I think I'm going to have to use QThread,
Workers, QRunnable, QThreadPool, or something of that ilk.