Make short break screen disappear after timeout
I made the short break screen hide itself after the timeout.  This is
made to be a break where the user looks away from the screen for a
period of time, to give their eyes a rest, but doesn't necessarily stand
up and walk away from the computer.  (This is in the spirit of the
20-20-20 rule:  Every 20 minutes, look at something 20 meters away for
20 seconds or more.)  As such, we can end the break as soon as it's
over, instead of waiting for the user to end the break.

There were a few other things involved here that I did as well:

- I added a test to the conditional junction, to see if a short break
  should come next, instead of the default long break.
- I customized the short break screen to be different from the long
  break screen.  I don't feel like we need a timer for the user to watch
  (they shouldn't be looking at the screen anyway), so I changed it to
  some simple text.  (I plan on customizing this later.)  I'm also not
  sure we need an end break button, so I took it off.  (Maybe this
  should be a "Skip break" button, but if I want that I'll deal with it
  later.)