WIP Adjust QLabel font size to fit text.
OK, so I feel a little dumb.  I took me a long time to figure out how to
get the height of the QLabel with the current text, current font, and
word wrapping set to 'on'.  To do so, use to following code:

        font = self.my_label.font()
        metrics = QFontMetrics(font)
        label_size = metrics.size(Qt.TextWordWrap, self.big_text_label.text())
        label_height = label_size.height()
        label_width = label_size.width()

The next steps I'm going to work on is to determine the height available
to me, and then to slowly adjust the font height from its maximum to a
size that makes the QLabel fit within the size that I want.

I'm hoping that part is a bit easier.  Wish me luck.