Fix negative timer values
This should actually be really rare in real life, where the break
spacing and break lengths are more reasonable values.  But I've gone
ahead and fixed it here by simply making the timeout 0 seconds if it's
calculated to be negative.  This throws off the timing a little bit, as
really the program should jump to the middle of the early break
notification (or maybe it should just jump to the late break
notification?) but I don't have a mechanism in place to do either of
those things, and where this is unlikely to happen in the first place,
I don't think I'm going to worry about it.