Add long break screen
Currently it processes the `break_ended` event on timer countdown
completion.  This should be done instead when the user presses the "Let
me get back to work" button.  The timeout at the end of the timer
countdown should either change the screen directly, or transition to a
`break_done` state, or something similar.  I'm kind of thinking the
latter, to keep as much of the logic in the state machine as possible.