Fix some timing issues
I did not clear out timers after I had set them.  This probably would've
been a non-issue most of the time, but it did come up in my very
short-duration timing that I'm using for testing, so I decided to change
it.  (Plus, it just feels sloppy not to.  I need to be more disciplined
about clearing potentially on-going timers in the future, when switching
states and whatnot.)  This should be solved now, but I'm going to keep
an eye on it.