Show next long break time in the logs
This is helpful for the user to see.  I would like to move this
information into a tool tip on the task bar icon in the future (or maybe
a relative version of it, i.e. "Next full break in 6 minutes" or
something similar).  It was also helpful to me to double check that it
was being set and modified correctly (especially after the timing issues
I was dealing with).