Set timings to sane values for actual use
Yay!  I feel this is at the point where I can use this myself (to
replace the janky shell script/Python program I was using earlier).

Eventually the values will be set in a configuration file and through a
configuration dialog, but for now, this is good enough to use on my
machine.

Whoo hoo!  (Just to reiterate.)  I've put quite a bit of effort in to
getting to something usable, and though it still needs polish, I'm glad
that I've gotten to this stage.  :-)