Include `stama` library as a git submodule
The `stama` library/module is a library I created that create state
machines.  (It is fairly well along, but also a work-in-progress.)  I
originally just had a symlink to the library and then used it like that.
In getting ready to release this, that's not going to work so well, so I
set up `stama` in it's own repository online and then "imported" it into
this project by creating it as a git submodule.