Remove context menu from glowbox
All of the options in the context menu were ones that I did want there.
There was one option to close/hide the glowbox, but which didn't do
anything to affect the state of the state machine, so it would cause
problems and was worse than useless.  The other two were test entries,
left over from when I was goofing around and learning.  They were both
useless (but not worse than useless, so a slight upgrade I guess?).

In the future I'd like to have context menu options, like the following:

- Skip the break
- Take the next regular break (if it's a notification for a short break)
- Take a short break (instead of a long break)
- Quit the program
- Configure the program
- Learn more about the program

A lot of these are options I'd like to add to the task bar icon as well.

Instead of implementing these in the `glowbox.py` module, I think I'll
create the context menu for the glowbox in the main program.  This makes
more sense to me as it makes the glowbox more composable (if I ever
wanted to use it in another context) and keeps creating both the task
bar icon context menu and glowbox context menu in the same place.