Get configuration values from a config file
... instead of hard-coding them into the program.

This still needs to account for missing values.  (I.e. what should a
variable be set to if it's not found in the config file?)  Currently,
every configuration option _must_ be present in the config file.  I want
some default values available if an option is not found in the config
file.  I plan to work on that next.

One note: I used `tomlkit` here instead of `tomllib` from the standard
library, because I plan on allowing the user to configure the program
from a graphical configuration window in the future.  I want to be able
to write those changes back to the file without losing any comments or
formatting.  The `tomlkit` modules allows that, whereas most other TOML
libraries do not.