Add default configuration
If no configuration files is found, or if values are missing from the
configuration file, then default values will be used.

I ended up creating the configuration as a dictionary with default
values, and then overriding those values if they were found in the
configuration file.  This was easier than setting the configuration
variables and then checking each one to see if it had an equivalent
value in the configuration file, and updating it if so.

Since my whole config is in a dictionary now, I think I may just pull
the values I need from the dictionary wherever it is that I need them,
instead of assigning the values to variables and then using them.  This
is my next step, to switch out those variables for the configuration
dictionary with the appropriate keys in place.