Stop logger from spewing debug stuff
I had forgotten that I had this on.

I need to implement proper file logging, with file rotation.  I also
need to create a `-v` command line option (and maybe a config file
option) so the user can set which level of logging they'd like to see on
the terminal.  These are both things I plan to do in the future.