Add optional "skip break" button to short breaks
This was a request from a tester.  I had originally thought tha short
break would be short enough that it didn't really need to be skipped if
you were already taking it.  I had considered (and will probably end up)
adding a context menu item to the glow box to skip the break that is
currently due.  When I get around to this, I may make the default option
for showing the "skip break" button false, since the user could just
skip the break from the glow box context menu.

Also, this sets whether the break screen has a "skip break" button on
initialization, meaning if the option is changed while the program is
running, the user would have to restart the program.  (This is a moot
point right now, since currently no configuration options can be changed
by the user once the program is started.  This is planned to change in
the future however.)  When I do get around to making the configuration
options configurable in a configuration screen, I will probably have to
look at this again and reconfigure the code.