Maybe fix the countdown clock glitch
... when first showing the long break screen.

So to figure out what was going on, I added some logging to the
`breakscreen` module.  After I had done this the problem seemed to go
away.  My best guess is that maybe the QTime object needed to be
accessed in some way before it was fully initialized?  I'm not quite
sure.  I can't count this bug as fully closed, but I'll be monitoring it
for a while, and if I don't see it repeat or I'm unable to replicate it,
I'll officially close it.  (I'll reopen it if it comes back.)

On that note: this bug is a regression.  I don't recall ever having this
problem before I created a base class for the break screens and then
inherited it into the separate long and short break screens.  I'm not
quite sure what to make of this.

Additionally I separated how logging was handled in the main program to
use two separate logging handlers, one for debugging on a SocketHandler,
and one for logging info to the console, so the user could have a sense
of what's happening.

(There's further work that will be done with logging.  This is not the
end.)