Add a tool tip to the task bar icon
It shows when upcoming breaks are scheduled for.

I would sometimes go to the logs to see when my next break will be.
Even then it wasn't alway straight forward as the short breaks would be
shown in relative time.  Now I can simply mouse over the task bar icon
to get that information.  (Also, it wouldn't be fair to the user to have
to check the logs for that information either.)

(I did however make the logs a bit more clear as well, now showing the
proper time on the clock for whatever breaks are next scheduled.)

I also refactored `next_long_break_clock_time` and
`next_short_break_clock_time` to `next_long_break_unix_time` and
`next_short_break_unix_time` to distinguish them a bit better from the
formatted clock time variables that I ended up using.

Also, on a sad note, the glitch in the countdown timer manifest itself.
I think I might've realized what's going on though:  It's using the
value of whatever was last there.  I feel a bit silly not realizing the
issue earlier and that I ought to reset the timer either on entrance or
exit, not just on initialization.  Derp.