Fix the countdown glitch for the long break screen
... wherein the countdown would start with the wrong countdown for the
first second.

OK.  Sometimes I can be slower than molasses, but only half as sweet.
I needed to update the label on showing the break screen.  It was that
simple.  (Before it would only update on a one-second tick.)  I
should've recoginized this earlier, but I've got it fixed now.

(I'm pretty certain this time.)