Fix some linting errors
Tons of fun!  /s

`pylint` is still raising quite a few issues, but I'll get to them when
I have the time.  It's still much better than it was before I did the
linting.

BTW, the headings changing from `################` to `# # # # # # # # `
is because flake8 complained about multiple `#` symbols in a row.
I sometimes use a double `#` to denote something I want to come back to,
so I'd still like to see that warning, but not for headings.  This
seemed like the simplest way to deal with that.