Add timeout length to short break screen text
This gives the user more information as to what is going on.  I've been
using this so long (and knew completely what was happening) that I'd
forgotten to think about the new user experience.  Being dropped into a
break screen with no indication of how long it's going to last -- that
is not good user experience.  This should be better.

(This was inspired partly by Matt, asking how to skip short breaks, and
me then realizing that the way the screen had been could be a bit
lacking in information.  I did also implement the skip button BTW.  See
commit `6f75ccaab` for more information.)