Adjust the long break screen
... to show a message about getting away from the computer.

I also adjusted the countdown timer to only show minutes and seconds.
You don't need a timer to tell you to step away from the computer for
hours at a time -- that's called 'life'.

I didn't realize it would be that easy to format the time, otherwise I
would've done it sooner.  (Some parts of QT seem a little
over-complicated to me sometimes.)