Get started with packaging
I `pip install`ed `nuitka` to try out as a compiler, and it mostly
worked!  (The one issue I had was that it could not find the icon, and
complained about it a bit, throwing an error, but everything else seemed
to work just fine.)  It's actually quite an amazing package.  In the
future, I will use this to compile Gentle Break Reminder into a binary
that's easier to install.

Here's the command I used to compile:

    nuitka3 --standalone --plugin-enable=pyside6 gentle.py