Get started with AFK status
This was actually a module that I created in another repo, playing
around and trying to figure out how to solve this problem.  (I suppose
this is bad practice, and I really ought to get in the habit of
developing in a feature branch, whether or not it shows up in the public
eye.  I'll try to be better about this.)

Python's `threading` module does not play well with QT (or at least
PySide6), so you need to use QT's own `QThread` class to do threading.
(I feel this is understandable as you are kind of working with both
Python in the `threading` module and C++ through QT.)  This requires
creating a worker object that you can then move to the QThread.  This
also requires communicating with the worker through signals and slots,
even for starting and stopping the worker, which need to be properly set
up.  Finally, everything needs to be shutdown properly;  if you leave
the worker or the thread running when shutting down the program, you'll
get some complaints and errors.

Next I plan on incorporating this into a state machine that includes a
limbo state for short bursts of activity seens when in and AFK state,
say for example if a user bumps their desk and moves the mouse, or
adjusts some playing music, or locks the screen.