WIP Try a different approach: add limbo to worker
I think I may be adding more complexity than is worth it here.  I don't
think that I need another layer built on top of what I've already done.
It may be better instead to improve what is already there.  I'm going to
try to add a limbo state to the `afk_worker` module.  Thinking about it,
the structure for a simple state machine based on switch/case statements
is already there.  Although I may do it inside out, "checking" for the
event first, since it's already implied in the callbacks, and then
checking for the state.

I also had a thought last night (while trying to sleep) that it may be a
good idea to run whatever custom events I want to be triggered by state
changes in the afk_worker, in the main thread.  Hopefully this would
help avoid problems with cross-thread issues.  To do this, I plan on
adding signals that activate slots in an object that stays in the main
thread.  I need to figure out a good way to test this.