Get a limbo state working
I need to return the time of last input and the time when the user
entered limbo as two relevant times, instead of when transitioning to
the AFK state and transitioning to the "at computer" state,
respecitvely.  Because of the timeout buffer for both of the latter,
they come after the user's actual last input activity or when they
actually came back to the computer.