Convert afk and at_computer actions to Signals
This should make it easier to use this across threads.

I still need to do the same for the limbo actions, and also convey
`last_input_time` and `last_akf_time`.

Also, I found the following video series useful:

https://www.youtube.com/playlist?list=PL6CJYn40gN6jgr-Rpl3J4XDQYhmUnxb-g

It deals with Qt in C++, but as long as you understand the concepts well
enough, you can apply it to PySide or PyQt as well.  (I just wanted to
put it here for reference, it case it became useful in the future.)