Move setting the next long break time
...to the right place: when the previous long break is finished.

I don't know why I didn't do this is in the first place.  Instead of
doing it this way I had tested to see if the previous long break time
had passed in the `has_short_break_before_long_break` function, and then
set the new break time if it had passed.  Thie did create some weird
behavior in testing, but because normal break times were further apart,
it never manifest in regular usage.  It makes much more sense to just
set the next long break time when the previous long break has finished.

I also need to add a timeout to break notifications (or something),
because I could see an issue where the user is away from the keyboard
for a while, comes back to a short break notification, clicks it, and
then when out of that, immediately gets a long break notification.  This
should be partially fixed by using input detection (or AFK status) which
I'm working on, but we still need reasonable settings if the user
disables input detections or if it's not available on the user's system.
(I think this last one may be the case in Wayland, for instance.)