Add a chime when the long break is finished
So the user can be notified that their long break is done, even if they
aren't in front of the computer (which hopefully they wouldn't be).

TODO: I need to track down the license and attribution for the sound
file I'm using.  I'm pretty sure I got it from a libre sound files
website, with a license that allows me to use it, but I don't quite
remember where from.  I'm going to leave it out of the repository until
I can get that done.

Also, I'd like to do something similar for a short break (with a
"smaller" sound) but I need to approach it a bit differently.  Simply
adding it to the break exit action will play the sound even if the user
is skipping the break (in which case they don't need to be notified that
the break is ended).

I've also added `singing_bowl.wav` to the `.gitignore` file for the
moment.  I'll stop ignoring it when I find a properly licensed file.