Allow scheduling events in an AFK period
This allows someone to schedule events a certain amount of time after an
AFK status has been met.  This will allow me, for instance, to consider
a break completed after the user has been AFK for a certain amount of
time (mostly likely equal to the amount of time that would've been spent
in the break, at least for long breaks).

I would've liked to been able to pass a dictionary of timeouts and
Signals, where the Signals would be emitted at the timeout.  However,
PySide Signals are required to be created on a class.  (They can't even
be created on an instance.  See https://stackoverflow.com/a/36561787 for
more info.)  This makes it impossible to create an independent signal in
the main thread, which I can then use as I please (to, for example,
affect a state machine).

One thing I need TODO in the future is to make it possible to change the
scheduled times, so that if a user changes their configuration while the
program is running, the program uses the updated config.  The
alternative is to require the user to restart the program, which I don't
want to do.  I actually need to do this for several things in the
program.  (I can't think of any off the top of my head, but I remember
needing to do this.)  One thing I need to be aware of in doing so is
race conditions.  What if I modify the dictionary of scheduled times and
functions, and then update the list of times in the AFKWorker, but
something happens in between those two processes?   Hmm... Fun!  For
this reason I may need to internalize the dictionary of scheduled
events, even though I was trying to move away from stuff like that.

(Ah, and I thought of another example:  When changing the color for the
glowbox, the color is set on initialization, so I either need to modify
that class to be able to update the color, or destroy the instance and
create it anew.  Bleh!  At least this shouldn't involve race conditions,
that I can think of.)