Update githook to provide better instructions
If the `requirements.txt` file doesn't match, it provides instructions
on how to uninstall unneeded modules (and only unneeded modules) and do
an updated install of all modules in `requirements.txt` (if they're not
already installed).

Previously, I was uninstalling everything in `requirements.txt` and
reinstalling everything, which was kind of silly.