Fix some formatting
I think I changed some of the formatting earlier to match `black`'s
formatting convention, but I didn't care for it.  I changed it back
(where I was able to without having to fight with `black`).

`black` likes to have either zero lines or two lines between a comment
and a function definition.  I use a comment in many cases as a heading
for a new section of code, preceded by two blank lines and followed by a
single blank line.  (Two lines provides white space separation from
other sections, which is good for new sections, but not for keeping a
heading with its section.)