Start refactoring various states to subclasses
I did this because I want it to be easier to "read" what each state is
doing.  This will hopefully make the code easier to read for any future
humans who read it.

I'd also like to (eventually) add functionality to the `stama` library
that can show the various states in either a state diagram/chart or in a
state table, with the actions included.  This will only be helpful if
the actions are simple (e.g. a function that summarizes what it does in
its name, instead of the whole contents of the function, i.e.
"show_short_break_late_notification()" instead of including all the
minutae of what that function does).  This is in preparation for that.

What spawned all of this was including the AFK functionality in the
`gentle` program, which didn't work when showing the short break
notifications.  (The notification would stay on screen, but the state
machine wouldn't react properly to clicks from the notification once it
was out of one of the showing notification states.)  I needed to add a
function that hid the notification whenever the notification was exited
(what should be a fairly simple task), but the action function and the
action assignment felt too disjointed.  So I've made the decision to
refactor into subclasses.

(Also, subclassing just "feels" better and less dirty than monkey
patching.  I don't know why I didn't do it originally.)