Refactor two more short break states
Also, in doing and testing this, I've discovered a bug.  When entering
an AFK state, anytime the computer receives input, even when going into
limbo, it will reset all scheduled events.  This is not what I want to
happen.  If the user is AFK for nine minutes, bumps their desk while
walking by and jolts the mouse, the programs should only wait one more
minute to determine that an AFK break has been completed, not the ten
more it currently would.  This is what I intend to work on next.