Move small functions to proper subclass method
So instead of having a function which only calls another function, I
just moved that inner function to the subclass method it belonged in.
We no longer need that outer function (because we don't have to assign
it to anything.  Clean!)  I also did this for functions that only had 1
or 2 lines to them as well.

I also cleaned up formatting a little bit.

I think I'm ready to test drive this for a bit and look for bugs, but
overall, it should be working a bit better (esp. AFK functionality).