Set up more events for testing
I'm going to run this for a while to see if I can pinpoint what is going
on.  In brief testing, everything works OK, but again, I only saw these
issues after running the program for about 24 hours.  I'm still not
quite sure what is going on.

Edited to add: I think I may know what is going on.  The calling
`singleShot()` on a timer instance creates multiple timers that will
each be called when they are complete.  This is why I'm getting timeouts
when I'm not excpecting them.  When I exit out of an AFK event of a
short duration into waiting for the next it creates a new timer that
persists, along with any previous timers created, thus I get multiple
timeout events when I'm not expecting them.

I've confirmed from the following links:

- https://stackoverflow.com/a/72703906
- https://www.qtcentre.org/threads/22832-multiple-QTimer-singleShot()-calls

I think the solution I'm going to pursue is to create two QTimer objects
that have the `singleShot` property: one for the long break timer and
one for the short break timer.  Then I can stop them as appropriate and
start them anew, overriding any previous time that the timer was set
for.

(I kind of think that my running the `stop()` method on `global_timer`
was pointless as well, as I was stopping a timer that had never really
been set up or run in the first place.  It was just spinning off other
timer instances.)