Fix time to next break showing as a float
The tool tip would show as something like "In less than 43.0½", which
was not good.  It now no longer shows the decimal part of the number.

On a related note, I may just show the time to the nearest minute, i.e.
"In about 43 minutes".  Also, I'd like to improve formatting a bit.