Make showing relative and clock times configurable
Having both the relative time ("In about ... minutes") and the clock
time in the tool tip was often too much information.  Sometimes it was
useful to have one or the other (or both) too.  So I've made it
configurable so that I can enable them both independently through the
configuration.