Change splash screen and make timeout configurable
My friend preferred a bit more color on the splash screen.

Also, instead of making the splash screen configuration true/false, it
now takes a timeout.  If the timeout is zero (or less) it doesn't show
the splash screen.  If it's more than zero, it shows it for that amount
of milliseconds.  This makes things more configurable.

I also adjusted how the default config values are set, so it's a bit
more obvious what is going on (hopefully).