Get the "About" window working right (mostly)
It was a stupid mistake.  I passed in the result of the
`about_window.show` function, instead of the function itself.  Doh!

I think stepping away from this and coming back to it with fresh eyes
helped me to spot the problem quickly.  (I've been working on several
other things lately.)

I did uncover a bug though:  When closing the "About" window using the
close button that is part of the window decoration, it will exit the
program entirely.  Not what I want.